requirements.txt
pywinauto==0.6.8
uiautomation==2.0.18
pytesseract==0.3.10
opencv-python==4.10.0.84
pydantic==2.8.2
typer==0.12.5
loguru==0.7.2
pyyaml==6.0.2
pytest==8.3.2
psutil==6.0.0
pyautogui==0.9.54
pillow==10.4.0
pygetwindow==0.0.9
rich==13.7.1
